rearrange tests
