+++ crates/burn-vision/src/filter.rs
@@ -1,5 +1,3 @@
-// The `Config` derive macro expands to `burn::...` paths, so it needs a crate
-// alias named `burn` in scope.
 use burn_core as burn;
 use burn_core::{
     Tensor,
@@ -30,8 +28,8 @@ use burn_std::{
 /// * `border` - The [padding mode](PadMode) applied to the input before the
 ///              convolution.
 ///
-/// The kernel may be any float dtype: it is cast to the image's
-/// dtype so the convolution operands agree.
+/// # Retuns
+/// The filtered batched images, same shape as the input.
 pub fn filter2d(images: Tensor<4>, kernel: Tensor<2>, border: PadMode) -> Tensor<4> {
     let [_, channels, _, _] = images.dims();
     let [kh, kw] = kernel.dims();
@@ -133,7 +131,6 @@ mod tests {
     use super::*;
     use burn_core::tensor::Tolerance;
 
-    /// A 1x1 identity kernel is a no-op regardless of border (no padding).
     #[test]
     fn filter2d_identity() {
         let images = Tensor::<4>::from([[[[1., 2.], [3., 4.]]]]);
@@ -144,40 +141,6 @@ mod tests {
             .assert_approx_eq(&out.to_data(), Tolerance::<f32>::balanced());
     }
 
-    /// A 3x3 box blur with a replicate (`Edge`) border on a constant image
-    /// returns the same constant (the weights sum to 1 and every neighbour
-    /// equals the centre).
-    #[test]
-    fn filter2d_box_blur_edge_constant_image() {
-        let images = Tensor::<4>::from([[[[5., 5., 5.], [5., 5., 5.], [5., 5., 5.]]]]);
-        let kernel = Tensor::<2>::from([[1. / 9.; 3]; 3]);
-        let out = filter2d(images.clone(), kernel, PadMode::Edge);
-        images
-            .to_data()
-            .assert_approx_eq(&out.to_data(), Tolerance::<f32>::balanced());
-    }
-
-    /// The same box blur with a zero (`Constant`) border darkens the edges: a
-    /// pixel's window loses the cells that fall outside the image. This
-    /// distinguishes the border mode from the `Edge` case above.
-    #[test]
-    fn filter2d_box_blur_constant_border() {
-        let images = Tensor::<4>::from([[[[5., 5., 5.], [5., 5., 5.], [5., 5., 5.]]]]);
-        let kernel = Tensor::<2>::from([[1. / 9.; 3]; 3]);
-        let out = filter2d(images, kernel, PadMode::Constant(0.));
-        // Real cells per 3x3 window (clamped to the 3x3 image), times 5/9.
-        let expected = Tensor::<4>::from([[[
-            [20. / 9., 30. / 9., 20. / 9.],
-            [30. / 9., 45. / 9., 30. / 9.],
-            [20. / 9., 30. / 9., 20. / 9.],
-        ]]]);
-        expected
-            .to_data()
-            .assert_approx_eq(&out.to_data(), Tolerance::<f32>::balanced());
-    }
-
-    /// A `[1, 3]` kernel filters only along width, leaving the height axis
-    /// untouched.
     #[test]
     fn filter2d_rectangular_kernel_width_only() {
         let images = Tensor::<4>::from([[[[1., 2., 3.]]]]);
@@ -190,8 +153,6 @@ mod tests {
             .assert_approx_eq(&out.to_data(), Tolerance::<f32>::balanced());
     }
 
-    /// Channels are filtered independently — a `1x1` scaling kernel scales each
-    /// channel on its own with no cross-channel mixing.
     #[test]
     fn filter2d_is_depthwise() {
         let images = Tensor::<4>::from([[[[1.]], [[10.]]]]); // [1, 2, 1, 1]
@@ -199,6 +160,39 @@ mod tests {
         let out = filter2d(images, kernel, PadMode::Reflect);
         let expected = Tensor::<4>::from([[[[2.]], [[20.]]]]);
         expected
+            .to_data()
+            .assert_approx_eq(&out.to_data(), Tolerance::<f32>::balanced());
+    }
+
+    #[test]
+    fn box_blur_constant_image_is_identity() {
+        let images = Tensor::<4>::from([[[[5., 5., 5.], [5., 5., 5.], [5., 5., 5.]]]]);
+        let blur = BoxBlurConfig::new().init(&Device::default());
+        let out = blur.apply(images.clone());
+        images
+            .to_data()
+            .assert_approx_eq(&out.to_data(), Tolerance::<f32>::balanced());
+    }
+
+    #[test]
+    fn box_blur_matches_manual_filter2d() {
+        let images = Tensor::<4>::from([[[[1., 2., 3.], [4., 5., 6.], [7., 8., 9.]]]]);
+        let kernel = Tensor::<2>::from([[1. / 9.; 3]; 3]);
+        let expected = filter2d(images.clone(), kernel, PadMode::Reflect);
+        let out = BoxBlurConfig::new().init(&Device::default()).apply(images);
+        expected
+            .to_data()
+            .assert_approx_eq(&out.to_data(), Tolerance::<f32>::balanced());
+    }
+
+    #[test]
+    fn box_blur_zero_probability_is_passthrough() {
+        let images = Tensor::<4>::from([[[[1., 2., 3.], [4., 5., 6.], [7., 8., 9.]]]]);
+        let blur = BoxBlurConfig::new()
+            .with_probability(0.0)
+            .init(&Device::default());
+        let out = blur.forward(images.clone());
+        images
             .to_data()
             .assert_approx_eq(&out.to_data(), Tolerance::<f32>::balanced());
     }
